Split chat service regression tests for structural gate
